Tile system first test
Procedurally generated tile system, with just one procedural mesh, one draw call. This is a very efficient system. We can go 10k tiles with that no problem.
I made a little test to see if the tiles are working properly, so there's a set timer inside the BP_Grid that just selects random tiles.
Later it will be necessary to make countries that are defined by the tiles. Once you click on a tile you should be able to identify and highlight the current country you are clicking on.